:recycle: css -> types
