+++ @navikt/core/react/src/layout/box/Box.tsx
@@ -1,12 +1,12 @@
 import cl from "clsx";
 import React, { forwardRef } from "react";
 import { OverridableComponent } from "../../util/OverridableComponent";
+import { getResponsiveProps } from "../utilities/css";
 import {
   SpaceDelimitedAttribute,
   ResponsiveProp,
   SpacingScale,
-  getResponsiveProps,
-} from "../utilities/css";
+} from "../utilities/types";
 import {
   BackgroundToken,
   BorderColorToken,
