+++ @navikt/core/react/src/layout/grid/HGrid.tsx
@@ -1,11 +1,7 @@
 import React, { forwardRef, HTMLAttributes } from "react";
 import cl from "clsx";
-import {
-  getResponsiveProps,
-  getResponsiveValue,
-  ResponsiveProp,
-  SpacingScale,
-} from "../utilities/css";
+import { getResponsiveProps, getResponsiveValue } from "../utilities/css";
+import { ResponsiveProp, SpacingScale } from "../utilities/types";
 
 export interface HGridProps extends HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
